Solved some bugs due to redesign and submodule commit correctly set
